Add model document for MetaKg; refactor models module
