+++ src/model/__init__.py
@@ -1,0 +1,4 @@
+from .metakg import MetaKGDoc
+from .smaprtapi import SmartAPIDoc
+
+__all__ = [MetaKGDoc, SmartAPIDoc]
